+++ src/abstract/service/auth.ts
@@ -1,9 +1,9 @@
-import { factory as SFR } from '@surfrock/soap-parser';
+import { SurfrockFactory } from '@surfrock/factory';
 import { OK } from 'http-status';
 import { Service } from '../service';
 
 export namespace factory {
-    export import purchaseNumberAuth = SFR.service.auth.purchaseNumberAuth;
+    export import purchaseNumberAuth = SurfrockFactory.service.auth.purchaseNumberAuth;
 }
 
 /**