+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - support eslint
 - update debug@4
 - update @types/node@22
+- reimplement using @surfrock/factory
 
 ### Deprecated